# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DailyProgressPoint": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "meal_count": {
            "title": "Meal Count",
            "type": "integer"
          },
          "totals": {
            "$ref": "#/components/schemas/NutrientTotals"
          }
        },
        "required": [
          "date",
          "meal_count",
          "totals"
        ],
        "title": "DailyProgressPoint",
        "type": "object"
      },
      "DailyProgressResponse": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "meal_count": {
            "title": "Meal Count",
            "type": "integer"
          },
          "totals": {
            "$ref": "#/components/schemas/NutrientTotals"
          }
        },
        "required": [
          "date",
          "meal_count",
          "totals"
        ],
        "title": "DailyProgressResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NutrientTotals": {
        "description": "Stored-meal nutrient totals aggregated for a reporting period.",
        "properties": {
          "calories": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Calories",
            "type": "string"
          },
          "carbohydrates_g": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Carbohydrates G",
            "type": "string"
          },
          "fat_g": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Fat G",
            "type": "string"
          },
          "fiber_g": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Fiber G",
            "type": "string"
          },
          "protein_g": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Protein G",
            "type": "string"
          }
        },
        "required": [
          "calories",
          "protein_g",
          "carbohydrates_g",
          "fat_g",
          "fiber_g"
        ],
        "title": "NutrientTotals",
        "type": "object"
      },
...
      "ProgressSummaryResponse": {
        "properties": {
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyProgressPoint"
            },
            "title": "Daily",
            "type": "array"
          },
          "daily_average": {
            "$ref": "#/components/schemas/NutrientTotals"
          },
          "days_with_meals": {
            "title": "Days With Meals",
            "type": "integer"
          },
          "meal_count": {
            "title": "Meal Count",
            "type": "integer"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/NutrientTotals"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "meal_count",
          "days_with_meals",
          "totals",
          "daily_average",
          "daily"
        ],
        "title": "ProgressSummaryResponse",
        "type": "object"
      },
...
      "TargetNutrientValues": {
        "properties": {
          "calories": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          },
          "carbohydrates_g": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbohydrates G"
          },
          "fat_g": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "fiber_g": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiber G"
          },
          "protein_g": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          }
        },
        "title": "TargetNutrientValues",
        "type": "object"
      },
...
      "TargetStatusResponse": {
        "description": "Neutral numeric comparison of today's stored consumption and configured targets.",
        "properties": {
          "consumed": {
            "$ref": "#/components/schemas/NutrientTotals"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "meal_count": {
            "title": "Meal Count",
            "type": "integer"
          },
          "percent_of_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionalNutrientValues"
              },
              {
                "type": "null"
              }
            ]
          },
          "remaining": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionalNutrientValues"
              },
              {
                "type": "null"
              }
            ]
          },
          "targets": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetNutrientValues"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "date",
          "meal_count",
          "consumed",
          "targets",
          "remaining",
          "percent_of_target"
        ],
        "title": "TargetStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeeklyDiagnosticsResponse": {
        "properties": {
          "days_in_period": {
            "title": "Days In Period",
            "type": "integer"
          },
          "days_with_logged_meals": {
            "title": "Days With Logged Meals",
            "type": "integer"
          },
          "days_with_scheduled_meals": {
            "title": "Days With Scheduled Meals",
            "type": "integer"
          },
          "logging_days_ratio": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Logging Days Ratio",
            "type": "string"
          },
          "meals_logged": {
            "title": "Meals Logged",
            "type": "integer"
          },
          "nutrition": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NutrientWeekMetric"
            },
            "title": "Nutrition",
            "type": "object"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "scheduled_meals": {
            "title": "Scheduled Meals",
            "type": "integer"
          },
          "target_comparison": {
            "$ref": "#/components/schemas/WeeklyTargetComparison"
          },
          "weight": {
            "$ref": "#/components/schemas/WeeklyWeightSummary"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "meals_logged",
          "days_with_logged_meals",
          "days_in_period",
          "logging_days_ratio",
          "scheduled_meals",
          "days_with_scheduled_meals",
          "nutrition",
          "weight",
          "target_comparison"
        ],
        "title": "WeeklyDiagnosticsResponse",
        "type": "object"
      },
      "WeeklyProgressResponse": {
        "properties": {
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyProgressPoint"
            },
            "maxItems": 7,
            "minItems": 7,
            "title": "Daily",
            "type": "array"
          },
          "meal_count": {
            "title": "Meal Count",
            "type": "integer"
          },
          "totals": {
            "$ref": "#/components/schemas/NutrientTotals"
          },
          "week_end": {
            "format": "date",
            "title": "Week End",
            "type": "string"
          },
          "week_start": {
            "format": "date",
            "title": "Week Start",
            "type": "string"
          }
        },
        "required": [
          "week_start",
          "week_end",
          "meal_count",
          "totals",
          "daily"
        ],
        "title": "WeeklyProgressResponse",
        "type": "object"
      },
...
      "WeightEntryList": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/WeightEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "entries",
          "limit",
          "offset"
        ],
        "title": "WeightEntryList",
        "type": "object"
      },
      "WeightEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "weight_kg": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight Kg",
            "type": "string"
          }
        },
        "required": [
          "weight_kg",
          "measured_at",
          "id",
          "created_at"
        ],
        "title": "WeightEntryResponse",
        "type": "object"
      },
      "WeightEntryWrite": {
        "additionalProperties": false,
        "properties": {
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "weight_kg": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 500.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "required": [
          "weight_kg",
          "measured_at"
        ],
        "title": "WeightEntryWrite",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Nutri-Box v1 integration API. Existing /api routes are the stable v1 contract.",
    "title": "Nutri-Box API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/progress/daily": {
      "get": {
        "operationId": "get_daily_progress_api_progress_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "maxLength": 64,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Daily Progress",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/progress/summary": {
      "get": {
        "operationId": "get_progress_summary_api_progress_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "maxLength": 64,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Progress Summary",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/progress/target-status": {
      "get": {
        "operationId": "get_today_target_status_api_progress_target_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "maxLength": 64,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Today Target Status",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/progress/today": {
      "get": {
        "operationId": "get_today_progress_api_progress_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "maxLength": 64,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Today Progress",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/progress/weekly": {
      "get": {
        "operationId": "get_weekly_progress_api_progress_weekly_get",
        "parameters": [
          {
            "in": "query",
            "name": "week_start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Week Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "maxLength": 64,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Weekly Progress",
        "tags": [
          "progress"
        ]
      }
    },
    "/api/progress/weekly-diagnostics": {
      "get": {
        "operationId": "weekly_diagnostics_api_progress_weekly_diagnostics_get",
        "parameters": [
          {
            "in": "query",
            "name": "week_start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Week Start",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyDiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Weekly Diagnostics",
        "tags": [
          "progress"
        ]
      }
    },
...
    "/api/weight-entries": {
      "get": {
        "operationId": "list_entries_api_weight_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "measured_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measured From"
            }
          },
          {
            "in": "query",
            "name": "measured_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Measured To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightEntryList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Entries",
        "tags": [
          "weight entries"
        ]
      },
      "post": {
        "operationId": "create_api_weight_entries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightEntryWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create",
        "tags": [
          "weight entries"
        ]
      }
    },
    "/api/weight-entries/{entry_id}": {
      "delete": {
        "operationId": "delete_api_weight_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete",
        "tags": [
          "weight entries"
        ]
      },
      "get": {
        "operationId": "get_api_weight_entries__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get",
        "tags": [
          "weight entries"
        ]
      },
      "put": {
        "operationId": "update_api_weight_entries__entry_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightEntryWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update",
        "tags": [
          "weight entries"
        ]
      }
    }
  }
}